# backend/app/api/routes/admin.py
"""Admin endpoints for imports and sync"""
from fastapi import APIRouter, Depends, HTTPException, UploadFile, File, Query
from pydantic import BaseModel
from sqlalchemy.orm import Session
import logging
import uuid

from app.db.database import get_db
from app.services.trade_import import TradeImportService
from app.services.daily_sync import run_daily_sync
from app.models.account import Account
from app.models.sleeve import Sleeve
from app.models.sync_log import SyncLog
from app.models.sheet_sync_config import SheetSyncConfig

router = APIRouter()
logger = logging.getLogger(__name__)


def _sleeve_for_client(db: Session, sleeve_id: str, client_id: str) -> Sleeve | None:
    """Return the sleeve only if it belongs to an account owned by the client."""
    return (
        db.query(Sleeve)
        .join(Account, Sleeve.account_id == Account.id)
        .filter(Sleeve.id == sleeve_id, Account.client_id == client_id)
        .first()
    )


# ---------------------------------------------------------------------------
# CSV import
# ---------------------------------------------------------------------------


@router.post("/import-historical")
def import_historical(
    client_id: str = Query(..., description="Client ID"),
    sleeve_id: str = Query(
        None,
        description="Optional. If set, import all rows into this sleeve. "
        "If omitted, auto-route each row by its Account + Strategy columns.",
    ),
    mode: str = Query("VALIDATE", description="VALIDATE or IMPORT"),
    file: UploadFile = File(..., description="CSV file with trades"),
    db: Session = Depends(get_db),
):
    """
    Import historical trades from a CSV file.

    CSV columns: Date, Symbol, Action, Quantity, Price, Commission, Strategy, Account, Notes

    Two routing modes:
    - **Single sleeve** (pass ``sleeve_id``): every row imports into that sleeve;
      the Account/Strategy columns are ignored.
    - **Auto-route** (omit ``sleeve_id``): each row is routed to the Account
      (by account number) + global Strategy (by name) named in the row, creating the
      account, strategy and/or sleeve if missing. Supports files spanning multiple
      accounts/strategies.

    Operation modes:
    - VALIDATE: validate only, write nothing (reports what *would* be created/imported)
    - IMPORT: validate and persist, with per-sleeve deduplication
    """
    if mode not in ["VALIDATE", "IMPORT"]:
        raise HTTPException(
            status_code=400, detail="Mode must be 'VALIDATE' or 'IMPORT'"
        )

    import_service = TradeImportService(db)
    file_content = file.file.read()

    try:
        if sleeve_id:
            sleeve = _sleeve_for_client(db, sleeve_id, client_id)
            if not sleeve:
                raise HTTPException(
                    status_code=404,
                    detail=f"Sleeve {sleeve_id} not found for client {client_id}",
                )
            result = import_service.import_from_file(file_content, sleeve_id, mode=mode)
        else:
            result = import_service.import_routed(file_content, client_id, mode=mode)

        if not result["success"]:
            raise HTTPException(
                status_code=400,
                detail={
                    "message": "Import validation failed",
                    "errors": result["errors"][:10],
                },
            )

        summary = {
            "total_rows": result["total"],
            "valid_trades": result["imported"],
            "duplicates_found": result["duplicates"],
            "errors": len(result["errors"]),
        }
        if not sleeve_id:
            summary["accounts_created"] = result["accounts_created"]
            summary["strategies_created"] = result["strategies_created"]
            summary["sleeves_created"] = result["sleeves_created"]

        response = {
            "status": "success",
            "mode": mode,
            "routing": "single-sleeve" if sleeve_id else "auto-route",
            "summary": summary,
        }

        if mode == "VALIDATE":
            response["message"] = "Validation passed, ready to import"
            if result["errors"]:
                response["validation_warnings"] = result["errors"][:10]
        else:
            response["message"] = f"Successfully imported {result['imported']} trades"
            response["sync_log_id"] = result["sync_log_id"]

            # Recompute weekly snapshots for all affected sleeves.
            try:
                from app.services.weekly_snapshot_service import WeeklySnapshotService
                svc = WeeklySnapshotService(db)
                affected = result.get("affected_sleeve_ids", [])
                if not affected and sleeve_id:
                    affected = [sleeve_id]
                for sid in affected:
                    svc.rebuild_sleeve(sid)
            except Exception:
                pass

        return response

    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Import failed: {e}")
        raise HTTPException(
            status_code=500, detail=f"Import processing failed: {str(e)}"
        )


# ---------------------------------------------------------------------------
# Manual Google Sheets sync
# ---------------------------------------------------------------------------


@router.post("/sync-daily")
def sync_daily(
    client_id: str = Query(..., description="Client ID"),
    sleeve_id: str = Query(..., description="Sleeve ID to sync trades into"),
    sheet_id: str = Query(..., description="Google Sheet ID"),
    range_name: str = Query(
        "Sheet1", description="Sheet range, e.g. Sheet1 or Sheet1!A:H"
    ),
    db: Session = Depends(get_db),
):
    """Sync trades from a Google Sheet into the specified sleeve."""
    sleeve = _sleeve_for_client(db, sleeve_id, client_id)
    if not sleeve:
        raise HTTPException(
            status_code=404,
            detail=f"Sleeve {sleeve_id} not found for client {client_id}",
        )

    try:
        result = run_daily_sync(
            db=db,
            sleeve_id=sleeve_id,
            sheet_id=sheet_id,
            range_name=range_name,
        )
    except Exception as e:
        logger.error(f"Daily sync failed: {e}")
        raise HTTPException(status_code=502, detail=f"Sync failed: {str(e)}")

    if not result["success"]:
        raise HTTPException(
            status_code=400,
            detail={"message": "Sync import failed", "errors": result["errors"][:10]},
        )

    # Recompute weekly snapshots for the synced sleeve.
    try:
        from app.services.weekly_snapshot_service import WeeklySnapshotService
        WeeklySnapshotService(db).rebuild_sleeve(sleeve_id)
    except Exception:
        pass

    return {
        "status": "success",
        "message": f"Synced {result['imported']} new trades from Google Sheets",
        "summary": {
            "total_rows": result["total"],
            "imported": result["imported"],
            "duplicates_skipped": result["duplicates"],
            "errors": len(result["errors"]),
        },
        "sync_log_id": result["sync_log_id"],
    }


# ---------------------------------------------------------------------------
# Sync logs
# ---------------------------------------------------------------------------


@router.get("/sync-logs")
def get_sync_logs(
    client_id: str = Query(...),
    skip: int = Query(0, ge=0),
    limit: int = Query(100, ge=1, le=1000),
    db: Session = Depends(get_db),
):
    """Get import/sync logs (all logs; client_id reserved for future filtering)."""
    try:
        total = db.query(SyncLog).count()
        logs = (
            db.query(SyncLog)
            .order_by(SyncLog.created_at.desc())
            .offset(skip)
            .limit(limit)
            .all()
        )
        return {
            "total": total,
            "logs": [
                {
                    "id": log.id,
                    "import_type": log.import_type,
                    "status": log.status,
                    "rows_processed": log.rows_processed,
                    "rows_success": log.rows_success,
                    "rows_failed": log.rows_failed,
                    "error_details": log.error_details,
                    "created_at": log.created_at,
                }
                for log in logs
            ],
        }
    except Exception as e:
        logger.error(f"Failed to fetch sync logs: {e}")
        raise HTTPException(status_code=500, detail="Failed to fetch logs")


# ---------------------------------------------------------------------------
# SheetSyncConfig CRUD — register/manage automated daily sync sources
# ---------------------------------------------------------------------------


class SyncConfigCreate(BaseModel):
    sleeve_id: str
    sheet_id: str
    range_name: str = "Sheet1"
    enabled: bool = True


class SyncConfigUpdate(BaseModel):
    enabled: bool


@router.post("/sync-configs", status_code=201)
def create_sync_config(body: SyncConfigCreate, db: Session = Depends(get_db)):
    """Register a Google Sheet → sleeve mapping for automated daily sync."""
    sleeve = db.query(Sleeve).filter(Sleeve.id == body.sleeve_id).first()
    if not sleeve:
        raise HTTPException(status_code=404, detail="Sleeve not found")

    cfg = SheetSyncConfig(
        id=str(uuid.uuid4()),
        sleeve_id=body.sleeve_id,
        sheet_id=body.sheet_id,
        range_name=body.range_name,
        enabled=body.enabled,
    )
    db.add(cfg)
    db.commit()
    db.refresh(cfg)
    return cfg


@router.get("/sync-configs")
def list_sync_configs(
    client_id: str = Query(None, description="Filter by client ID"),
    db: Session = Depends(get_db),
):
    """List all sheet sync configurations, optionally filtered by client."""
    q = db.query(SheetSyncConfig)
    if client_id:
        q = (
            q.join(Sleeve, SheetSyncConfig.sleeve_id == Sleeve.id)
            .join(Account, Sleeve.account_id == Account.id)
            .filter(Account.client_id == client_id)
        )
    return q.order_by(SheetSyncConfig.created_at.desc()).all()


@router.patch("/sync-configs/{config_id}")
def update_sync_config(
    config_id: str,
    body: SyncConfigUpdate,
    db: Session = Depends(get_db),
):
    """Enable or disable an automated sync config."""
    cfg = db.query(SheetSyncConfig).filter(SheetSyncConfig.id == config_id).first()
    if not cfg:
        raise HTTPException(status_code=404, detail="Sync config not found")
    cfg.enabled = body.enabled
    db.commit()
    db.refresh(cfg)
    return cfg


@router.delete("/sync-configs/{config_id}", status_code=204)
def delete_sync_config(config_id: str, db: Session = Depends(get_db)):
    """Remove a sync config."""
    cfg = db.query(SheetSyncConfig).filter(SheetSyncConfig.id == config_id).first()
    if not cfg:
        raise HTTPException(status_code=404, detail="Sync config not found")
    db.delete(cfg)
    db.commit()


# ---------------------------------------------------------------------------
# Weekly snapshot rebuild
# ---------------------------------------------------------------------------


@router.post("/rebuild-snapshots")
def rebuild_snapshots(
    sleeve_id: str = Query(None, description="Rebuild a single sleeve; omit for all sleeves"),
    db: Session = Depends(get_db),
):
    """Rebuild weekly portfolio snapshots.

    Recomputes all ``weekly_portfolio_snapshots`` rows from scratch.
    Pass ``sleeve_id`` to rebuild only one sleeve; omit to rebuild every sleeve.
    This is idempotent — safe to run after correcting historical trade data.
    """
    from app.services.weekly_snapshot_service import WeeklySnapshotService
    svc = WeeklySnapshotService(db)
    try:
        if sleeve_id:
            count = svc.rebuild_sleeve(sleeve_id)
            return {"status": "success", "sleeves_rebuilt": 1, "snapshots_written": count}
        else:
            results = svc.rebuild_all()
            return {
                "status": "success",
                "sleeves_rebuilt": len(results),
                "snapshots_written": sum(results.values()),
            }
    except Exception as e:
        logger.error(f"Snapshot rebuild failed: {e}")
        raise HTTPException(status_code=500, detail=f"Rebuild failed: {str(e)}")