# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Management System API
  description: |
    Multi-client, multi-strategy investment portfolio management backend.

    **Hierarchy:** Client → Account → Sleeve (Account × Strategy join) → Trades/Positions

    - A **Strategy** is a firm-wide definition (e.g. "Growth", "Income") — global and reusable.
    - A **Sleeve** is one strategy as run in one account — it owns trades, positions, and sync configs.
    - All resource IDs are UUIDs (string).
  version: 1.0.0
  contact:
    email: venkatasudheerc@gmail.com
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/admin/rebuild-snapshots:
    post:
      tags: [admin]
      summary: Rebuild weekly portfolio snapshots
      description: |
        Recomputes all `weekly_portfolio_snapshots` from scratch.
        Pass `sleeve_id` to rebuild only one sleeve; omit to rebuild all sleeves.
        Safe to run after correcting historical trade data.
      operationId: rebuildSnapshots
      parameters:
        - name: sleeve_id
          in: query
          required: false
          schema:
            type: string
          description: Optional. Rebuild only this sleeve.
      responses:
        "200":
          description: Rebuild result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  sleeves_rebuilt:
                    type: integer
                  snapshots_written:
                    type: integer
        "500":
          description: Internal error during rebuild

  # ─── Trades ─────────────────────────────────────────────────────────────────
...
  /api/v1/admin/import-historical:
    post:
      tags: [admin]
      summary: Import historical trades from a CSV file
      description: |
        CSV columns: Date, Symbol, Action, Quantity, Price, Commission, Strategy, Account, Notes.

        **Routing modes:**
        - `sleeve_id` provided → all rows go into that single sleeve
        - `sleeve_id` omitted → auto-route by Account + Strategy columns (creates accounts/strategies/sleeves as needed)

        **Operation modes:**
        - `VALIDATE` → dry-run; returns warnings without writing to DB
        - `IMPORT` → writes trades; skips duplicates
      operationId: importHistorical
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: Target client UUID
        - name: sleeve_id
          in: query
          required: false
          schema:
            type: string
          description: Target sleeve UUID (omit to enable auto-routing)
        - name: mode
          in: query
          required: true
          schema:
            type: string
            enum: [VALIDATE, IMPORT]
          description: VALIDATE for dry-run, IMPORT to persist
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file to import
      responses:
        "200":
          description: Import result (or validation result in VALIDATE mode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/sync-daily:
    post:
      tags: [admin]
      summary: Sync trades from a Google Sheet into a sleeve
      description: Fetches the public Google Sheet as CSV and imports new rows. Duplicate rows are skipped via `google_sheet_row_id`.
      operationId: syncDaily
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: sleeve_id
          in: query
          required: true
          schema:
            type: string
        - name: sheet_id
          in: query
          required: true
          schema:
            type: string
          description: Google Sheets document ID
        - name: range_name
          in: query
          required: false
          schema:
            type: string
            default: Sheet1
          description: Sheet tab / named range
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncDailyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/sync-logs:
    get:
      tags: [admin]
      summary: Get import/sync audit logs
      operationId: getSyncLogs
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: Client UUID (reserved for future filtering)
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated sync logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncLogsResponse"
  /api/v1/admin/sync-configs:
    post:
      tags: [admin]
      summary: Register a Google Sheet → sleeve mapping for automated daily sync
      operationId: createSyncConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncConfigCreate"
      responses:
        "201":
          description: Sync config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetSyncConfig"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [admin]
      summary: List all sheet sync configurations
      operationId: listSyncConfigs
      parameters:
        - name: client_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by client UUID
      responses:
        "200":
          description: List of sync configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SheetSyncConfig"
  /api/v1/admin/sync-configs/{config_id}:
    patch:
      tags: [admin]
      summary: Enable or disable an automated sync config
      operationId: updateSyncConfig
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncConfigUpdate"
      responses:
        "200":
          description: Updated sync config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetSyncConfig"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [admin]
      summary: Remove a sync config
      operationId: deleteSyncConfig
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"

# ─── Components ───────────────────────────────────────────────────────────────
components:
  parameters:
...
    skip:
      name: skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Pagination offset
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of results to return
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request (e.g. invalid CSV format)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items: {}
                    msg:
                      type: string
                    type:
                      type: string
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ── Clients ──────────────────────────────────────────────────────────────
...
    ImportResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        mode:
          type: string
          enum: [VALIDATE, IMPORT]
        routing:
          type: string
          enum: [single-sleeve, auto-route]
        summary:
          type: object
          properties:
            total_rows:
              type: integer
            valid_trades:
              type: integer
            duplicates_found:
              type: integer
            errors:
              type: integer
            accounts_created:
              type: integer
              description: Only present in auto-route mode
            strategies_created:
              type: integer
              description: Only present in auto-route mode
            sleeves_created:
              type: integer
              description: Only present in auto-route mode
        message:
          type: string
        validation_warnings:
          type: array
          nullable: true
          items:
            type: object
          description: Populated in VALIDATE mode
        sync_log_id:
          type: string
          nullable: true
          description: Populated in IMPORT mode
    SyncDailyResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
        summary:
          type: object
          properties:
            total_rows:
              type: integer
            imported:
              type: integer
            duplicates_skipped:
              type: integer
            errors:
              type: integer
        sync_log_id:
          type: string
    SyncLogEntry:
      type: object
      properties:
        id:
          type: string
        import_type:
          type: string
        status:
          type: string
        rows_processed:
          type: integer
        rows_success:
          type: integer
        rows_failed:
          type: integer
        error_details:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    SyncLogsResponse:
      type: object
      properties:
        total:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/SyncLogEntry"
    SyncConfigCreate:
      type: object
      required: [sleeve_id, sheet_id]
      properties:
        sleeve_id:
          type: string
        sheet_id:
          type: string
          description: Google Sheets document ID
        range_name:
          type: string
          default: Sheet1
        enabled:
          type: boolean
          default: true
    SyncConfigUpdate:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    SheetSyncConfig:
      type: object
      properties:
        id:
          type: string
        sleeve_id:
          type: string
        sheet_id:
          type: string
        range_name:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time